fix: 이미 소프트 삭제된 리뷰 물리 삭제시 평점을 두번 차감하는 버그 수정

+++ src/main/java/com/sbproject/deokhugam/review/service/Impl/ReviewServiceImpl.java
@@ -138,7 +138,9 @@ public class ReviewServiceImpl implements ReviewService {
 			throw new IllegalStateException("해당 리뷰를 삭제할 권한이 없습니다.");
 		}
 
-		review.getBook().deleteReviewRating(review.getRating());
+		if(review.getDeletedAt() == null) {
+			review.getBook().deleteReviewRating(review.getRating());
+		}
 
 		reviewRepository.delete(review);
 	}
